requirements.txt
# Aviation Operations Dashboard - Dependencies
# Phase 1: Foundation Setup

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
waitress>=2.1.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.9

# AIMS SOAP API Integration
zeep>=4.2.1
lxml>=4.9.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
pdfplumber>=0.10.0

# Scheduling
APScheduler>=3.10.0

# Environment & Configuration
python-dotenv>=1.0.0

# Date/Time Handling
pytz>=2024.1

# Logging
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
watchdog>=3.0.0

# Security
Flask-Limiter>=3.5.0
marshmallow>=3.20.0
bleach>=6.1.0
